# api/openapi.yaml
openapi: 3.1.0
info:
  title: ansible-automation-content
  version: '1'
  description: |
    Content management API for Ansible Automation Portal.

    Discovers, describes and resolves Ansible automation content — execution
    environments, collections, and the modules, plugins, roles and Event-Driven Ansible
    plugins inside them — across any OCI-compliant registry.

    ## Two levels of content

    The model distinguishes two granularities, and the distinction runs through every
    endpoint:

    - **Artifacts** are distributable units: an execution environment, a collection, a
      skill. They are pushed, signed, governed and promoted. They carry trust.
    - **Content items** are what lives inside an artifact: modules, filter/lookup/test
      plugins, roles, playbooks, EDA plugins. They are documented and searched, and have
      no independent lifecycle. Trust is inherited from the containing artifact.

    `/content` addresses artifacts, `/content-items` addresses content items. Conflating
    them is the most common modelling error in this domain.

    Both are **type-agnostic**. Onboarding a new content type — skills, node types,
    workflow definitions — is registering an adapter, not adding an endpoint.
    `/execution-environments` and `/collections` remain as thin aliases over `/content`
    so existing callers keep working; they are filters, not separate implementations.

    ## Designed for agents as well as humans

...

...
      tell which they have.
...

...

...

...
servers:
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - JWT: []
paths:
...
  /registries/{name}/sync:
    post:
      operationId: syncRegistry
      tags: [Registries]
      summary: Trigger discovery for one registry
      description: Deliberate and manual, matching the existing collections pattern, to avoid unbounded syncing.
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Sync started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  discovered: { type: integer }
        '404':
          description: No such registry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /content:
    get:
      operationId: listContent
      tags: [Content]
      summary: List content artifacts of any type
      parameters:
        - name: type
          in: query
          description: 'e.g. `execution-environment`, `collection`, or `oci-image` for unrecognised images.'
          schema: { type: string }
        - name: registry
          in: query
          schema: { type: string }
        - name: collection
          in: query
          description: Only artifacts shipping this collection, e.g. `cisco.ios`.
          schema: { type: string }
        - name: contentsKnown
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Matching artifacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ExecutionEnvironment' }
  /content/{ref}:
    get:
      operationId: getContent
      tags: [Content]
      summary: Get one content artifact
      parameters:
        - name: ref
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The artifact
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionEnvironment' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /content-items:
    get:
      operationId: searchContentItemsTypeAgnostic
      tags: [Content Items]
      summary: Search content items across all artifacts
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: type
          in: query
          description: 'module, filter, lookup, test, role, playbook, event_source, …'
          schema: { type: string }
        - name: collection
          in: query
          schema: { type: string }
        - name: in
          in: query
          description: Restrict to one artifact, by ref.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        '200':
          description: Matching items
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalItems: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ContentItemSummary' }
  /content-items/{fqcn}:
    get:
      operationId: getContentItemTypeAgnostic
      tags: [Content Items]
      summary: Full documentation for one content item
      description: |
        Scope with `in` to resolve within a specific artifact. Unscoped, and where the
        name exists in more than one, a `variants` array is returned rather than one
        being chosen — picking silently would produce an answer that looks authoritative
        and may describe a different collection version than the caller will run.
      parameters:
        - name: fqcn
          in: path
          required: true
          schema: { type: string }
        - name: in
          in: query
          description: Artifact ref to resolve within.
          schema: { type: string }
      responses:
        '200':
          description: The item, or its variants
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ContentItem'
                  - type: object
                    properties:
                      fqcn: { type: string }
                      variantCount: { type: integer }
                      variants:
                        type: array
                        items: { $ref: '#/components/schemas/ContentItem' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /execution-environments:
    get:
      operationId: listExecutionEnvironments
      tags: [Execution Environments]
      summary: List execution environments
      parameters:
        - name: registry
          in: query
          schema: { type: string }
        - name: collection
          in: query
          description: Only environments shipping this collection, e.g. `cisco.ios`.
          schema: { type: string }
        - name: contentsKnown
          in: query
          description: Filter by whether the contents could be determined.
          schema: { type: boolean }
      responses:
        '200':
          description: Matching environments
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ExecutionEnvironment' }
...
  /collections:
    get:
      operationId: listCollections
      tags: [Collections]
      summary: List collections available across all registries
      parameters:
        - name: q
          in: query
          description: Substring match on namespace, name or description.
          schema: { type: string }
      responses:
        '200':
          description: Matching collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Collection' }
  /collections/{namespace}/{name}:
    get:
      operationId: getCollection
      tags: [Collections]
      summary: Get a collection, including which environments provide it
      parameters:
        - name: namespace
          in: path
          required: true
          schema: { type: string }
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Collection' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /plugins:
    get:
      operationId: searchContentItems
      tags: [Content Items]
      summary: Search modules, plugins, roles and EDA plugins
      description: |
        The primary agent-grounding endpoint. Answers "does this environment provide the
        module I am about to generate a task for, and what are its real arguments?"
        without pulling an image.
      parameters:
        - name: q
          in: query
          description: Substring match on fully qualified name or short description.
          schema: { type: string }
        - name: type
          in: query
          description: 'Item type: module, filter, lookup, test, connection, event_source, …'
          schema: { type: string }
        - name: collection
          in: query
          schema: { type: string }
        - name: executionEnvironment
          in: query
          description: Restrict to items available in one environment.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        '200':
          description: Matching items
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalItems: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ContentItemSummary' }
  /plugins/{fqcn}:
    get:
      operationId: getContentItem
      tags: [Content Items]
      summary: Full documentation for one content item
      description: |
        Returns the complete argument specification including nested suboptions, plus
        examples and return values where the collection documents them. This is what an
        agent needs to verify generated syntax against the environment that will run it.
      parameters:
        - name: fqcn
          in: path
          required: true
          description: Fully qualified collection name, e.g. `cisco.ios.ios_vlans`.
          schema: { type: string }
        - name: executionEnvironment
          in: query
          description: |
            Resolve within a specific environment. Different environments may ship
            different versions of the same collection, so the documentation differs.
          schema: { type: string }
      responses:
        '200':
          description: The item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentItem' }
        '404':
          description: Not found in any known environment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /config/export:
    get:
      operationId: exportConfig
      tags: [Configuration]
      summary: Export the live content configuration as a declarative document
      description: |
        Emits configuration suitable for committing to Git. Secrets are never included;
        credential fields are emitted as environment-variable references so the exported
        document is safe to store in version control.
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, yaml]
            default: json
      responses:
        '200':
          description: The configuration
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentConfig' }
            application/yaml:
              schema:
                type: string
...